test: move benchmark tests to extension-vertx(need jdk17)
