+++ benchmark/pom.xml
@@ -28,11 +28,6 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>com.alibaba.fastjson2</groupId>
-            <artifactId>fastjson2-extension-vertx</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
             <groupId>com.alibaba</groupId>
             <artifactId>fastjson</artifactId>
             <version>${fastjson1x.version}</version>
@@ -95,11 +90,6 @@
             <groupId>io.github.wycst</groupId>
             <artifactId>wast</artifactId>
             <version>0.0.28</version>
-        </dependency>
-        <dependency>
-            <groupId>io.vertx</groupId>
-            <artifactId>vertx-core</artifactId>
-            <version>${vertx5.version}</version>
         </dependency>
         <dependency>
             <groupId>javax.annotation</groupId>
