+++ extension-vertx/src/main/resources/small_bench.json
