+++ extension-vertx/src/main/resources/wide_bench.json
